requirements.txt
customtkinter==5.1.3
pillow-heif==1.2.1
Pillow==10.0.1
streamlit==1.54.0